requirements.txt
aiofiles==0.8.0
aiohttp==3.8.3
aiosignal==1.2.0
aiosqlite==0.17.0
argcomplete==2.0.0
async-generator==1.10
async-timeout==4.0.2
asyncio-extras==1.3.2
asyncpraw==7.6.0
asyncprawcore==2.3.0
attrs==22.1.0
backports.cached-property==1.0.1
beautifulsoup4==4.11.1
Brotli==1.0.9
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-default-group==1.2.2
cloup==0.7.1
colorama==0.4.4
colour==0.1.5
commonmark==0.9.1
decorator==5.1.1
discord.py==2.0.1
distlib==0.3.4
filelock==3.6.0
frozenlist==1.3.1
glcontext==2.3.4
idna==3.3
isosurfaces==0.1.0
ManimPango==0.4.0.post2
mapbox-earcut==0.12.11
moderngl==5.6.4
moderngl-window==2.4.1
multidict==6.0.2
multipledispatch==0.6.0
mutagen==1.45.1
networkx==2.6.3
numpy==1.22.2
packaging==21.3
Pillow==9.0.1
pipenv==2022.1.8
platformdirs==2.5.1
praw==7.6.0
prawcore==2.3.0
pycairo==1.20.1
pycryptodomex==3.14.1
pydub==0.25.1
pygame==2.1.2
pyglet==1.5.21
Pygments==2.11.2
pyparsing==3.0.7
pyrr==0.10.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.27.1
rich==11.2.0
scipy==1.8.0
screeninfo==0.8
six==1.16.0
skia-pathops==0.7.2
soupsieve==2.3.2.post1
srt==3.5.1
sty==1.0.4
tqdm==4.62.3
typing_extensions==4.4.0
update-checker==0.18.0
urllib3==1.26.8
userpath==1.8.0
virtualenv==20.13.2
virtualenv-clone==0.5.7
watchdog==2.1.6
websocket-client==1.4.1
websockets==10.2
yarl==1.8.1
yt-dlp==2022.3.8.2
